requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

github-action-utils==1.1.0
PyGithub==2.3.0
